+++ lib/Executor/Benchmark/LocalExecutor.php
@@ -31,6 +31,17 @@ class LocalExecutor implements BenchmarkExecutorInterface
 
     public function execute(ExecutionContext $context, Config $config): ExecutionResults
     {
+        try {
+            return $this->doExecute($context, $config);
+        } catch (\Exception $e) {
+            throw new ExecutionError(
+                sprintf("Exception encountered in benchmark: %s\n\n[%s]\n\n%s", $e->getMessage(), get_class($e), $e->getTraceAsString()),
+            );
+        }
+    }
+
+    private function doExecute(ExecutionContext $context, Config $config): ExecutionResults
+    {
         if ($this->bootstrap) {
             require_once($this->bootstrap);
         }