+++ tests/Benchmark/Expression/ParserBench.php
@@ -41,7 +41,6 @@ class ParserBench
      */
     public function benchEvaluate(array $params): void
     {
-        usleep(10000);
         $this->parser->parse($this->lexer->lex($params['expr']));
     }